Do not leak callbacks
A session is identified by the variable 'uuid'. Multiple command
requests/responses are possible, hence the callback was registered only
once before. These were however not cleaned up.

Take a different approach, always register an event handler for the
session and check in the callback whether the response matched the
request. This results in a number of calls that grows linearly to the
number of open requests, but this number is usually small anyway.

Fixes https://github.com/englercj/node-esl/issues/54